[Feat] intergrate multiple settings controller into one
